Add a new constructor FileListener
This new constructor can be used in conjunction with the new SyscallConn
method added in commit 805d8b8.

This allows processes to create a SCTP listener, pass the underlying fd
to another process, and let that other process recreate the SCTPListener
struct from that fd.

The implementation is roughly based on `net.FileListener`, except it
returns a concrete SCTPListener instead of a `net.Listener`. The fd is
dup'd to ensure that the SCTPListener won't be affected when the
*os.File passed to FileListener is closed.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>